Tweaked referencing comment format for MDW and added [[ x64 ]] markers for searching. Reverted warning fixes in lnxmplayer.* (the warnings are caused by issues with string ownership, so better to have the compiler complain for now). Changed MCS_getprocessor() on Linux to return based on compile time target machine rather than uname (in case uname is inconsistent). Reverted changes to MCS_getprocessor() for other platforms (better done on a separate branch). Used MAP_FAILED constant to check mmap() error, rather than casting.

+++ engine/src/osxdce.cpp
@@ -328,7 +328,7 @@ Boolean MCScreenDC::getmouseclick(uint2 button, Boolean& r_abort)
 		r_abort = wait(0.0, False, True);
 		if (r_abort)
 			return False;
-		return MCEventQueueGetMouseClick((uint32_t)button);
+		return MCEventQueueGetMouseClick(button);
 	}
 	
 	r_abort = this->wait(0.0, False, True);
